Moved files into the util directory.
- Moved the files KeepTablesOnly.java and Sort_merge3.java into the
  util directory as this seems like a more appropriate location
  for the files.
- Updated any references related to these files.
