+++ code/componentsrunner/src/main/java/ca/sfu/cs/componentsrunner/app/RunComponent.java
@@ -8,7 +8,7 @@ import java.text.MessageFormat;
 import com.mysql.jdbc.Connection;
 
 import ca.sfu.cs.common.Configuration.Config;
-import ca.sfu.cs.factorbase.tables.Sort_merge3;
+import ca.sfu.cs.factorbase.util.Sort_merge3;
 
 
 public class RunComponent {
