+++ code/factorbase/src/main/java/ca/sfu/cs/factorbase/util/KeepTablesOnly.java
@@ -1,4 +1,4 @@
-package ca.sfu.cs.factorbase.tables;
+package ca.sfu.cs.factorbase.util;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
